client: enable keepalive on remote connection
copy_bidirectional() won't return if an used connection with remote has been
disconnected because of a network problem, resulting in the
tunnel no longer working once all the connections are in this state.

Fix this by enabling TCP keepalive with a shorter timeout that the OS
default so those connections will timeout and we'll try to reconnect
then.

See https://github.com/tokio-rs/tokio/issues/4556 for details.