server: clean up timed out connections on timer
If the client is disconnected for a while because of the network, it
will try to re-establish new connections to the server once the network
is back. But those new connections could be rejected if the server
reached its connection limit with the old disconnected connections.

Fix this by purging those old connections if new ones have not been
established for a while.

Fix #36